Some changes

+++ Algorithms/Algebraic/Power.hpp
@@ -1,12 +1,14 @@
 #ifndef D74F22B2_4797_4216_83CC_7768B6809CB3
 #define D74F22B2_4797_4216_83CC_7768B6809CB3
 
+#include <cstdint>
+
 namespace Algorithms {
   class Power {
   public:
-    long double PowerRecursion(long double value, int n);
-    long double PowerIterative(long double value, int n);
-    long double PowerOptimized(long double value, int n);
+    long double PowerRecursion(long double value, int64_t n);
+    long double PowerIterative(long double value, int64_t n);
+    long double PowerOptimized(long double value, int64_t n);
   };
 }
